tweaked music installation

+++ Quick_BGM_Install.sh
@@ -64,8 +64,12 @@ install_bgm_1() {
 minimum=0
 clear
 prep_work
-gdown https://drive.google.com/uc?id=1hv2nXThZ5S4OkY-oLGKwMtjmfRYy2cFe -O $HOME/retropie_music_overlay/bgm.zip
-unzip -foq $HOME/retropie_music_overlay/bgm.zip -d $HOME/RetroPie
+if [ -f "/home/pi/RetroPie/roms/music/arcade/arcade81.mp3" ]; then
+	echo "Found Music!"
+else
+	gdown https://drive.google.com/uc?id=1hv2nXThZ5S4OkY-oLGKwMtjmfRYy2cFe -O $HOME/retropie_music_overlay/bgm.zip
+	unzip -foq $HOME/retropie_music_overlay/bgm.zip -d $HOME/RetroPie
+fi
 setup
 rebootq
 exit
@@ -74,11 +78,20 @@ install_bgm_2() {
 minimum=0
 clear
 prep_work
-gdown https://drive.google.com/uc?id=1hv2nXThZ5S4OkY-oLGKwMtjmfRYy2cFe -O $HOME/retropie_music_overlay/bgm.zip
-gdown https://drive.google.com/uc?id=1-BHwb4oT6GiwpRv7l3VLHuJLsRxScGNV -O $HOME/retropie_music_overlay/custombgm.zip
-unzip -foq $HOME/retropie_music_overlay/bgm.zip -d $HOME/RetroPie
-unzip -foq $HOME/retropie_music_overlay/custombgm.zip -d $HOME/RetroPie
-rm -f $MUSIC_DIR/custom/'No Music in Folder.mp3'
+if [ -f "/home/pi/RetroPie/roms/music/arcade/arcade81.mp3" ]; then
+	echo "Found Music!"
+else
+	gdown https://drive.google.com/uc?id=1hv2nXThZ5S4OkY-oLGKwMtjmfRYy2cFe -O $HOME/retropie_music_overlay/bgm.zip
+	unzip -foq $HOME/retropie_music_overlay/bgm.zip -d $HOME/RetroPie
+fi
+if [ -f "/home/pi/RetroPie/roms/music/custom/3 Inches Of Blood- Deadly Sinners.mp3" ]; then
+	echo "Found Music!"
+else
+	gdown https://drive.google.com/uc?id=1-BHwb4oT6GiwpRv7l3VLHuJLsRxScGNV -O $HOME/retropie_music_overlay/custombgm.zip
+	unzip -foq $HOME/retropie_music_overlay/bgm.zip -d $HOME/RetroPie
+	unzip -foq $HOME/retropie_music_overlay/custombgm.zip -d $HOME/RetroPie
+	rm -f $MUSIC_DIR/custom/'No Music in Folder.mp3'
+fi
 setup
 rebootq
 exit
